scripts/appconfig/requirements.txt
azure-identity==1.21.0
python-dotenv==1.1.0
azure-appconfiguration==1.7.1